+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Add `Debug` impls for `Collector`, `Handle`, and `guard`.
+
 ## [0.4.0] - 2018-02-10
 ### Changed
 - Update dependencies.